+++ src/main/java/com/bulka/service/AuthorService.java
@@ -31,6 +31,9 @@ public class AuthorService {
 
     public AuthorDTO getAuthorById(Long id) {
         Author author = authorRepository.getAuthorById(id);
+        if (author == null) {
+            return null; // или бросить исключение
+        }
         AuthorDTO authorDTO = new AuthorDTO();
         authorDTO.setId(author.getId());
         authorDTO.setName(author.getName());
@@ -39,21 +42,31 @@ public class AuthorService {
     }
 
     public void createAuthor(AuthorDTO authorDTO) {
-        Author author = new Author();
-        author.setName(authorDTO.getName());
-        author.setSurname(authorDTO.getSurname());
-        authorRepository.createAuthor(author);
+        Author author = authorRepository.getAuthorByNameAndSurname(authorDTO.getName(), authorDTO.getSurname());
+        if (author!= null) {
+            throw new RuntimeException("Автор с таким же именем и фамилией уже существует"); // или бросить исключение
+        }
+        Author newAuthor = new Author();
+        newAuthor.setName(authorDTO.getName());
+        newAuthor.setSurname(authorDTO.getSurname());
+        authorRepository.createAuthor(newAuthor);
     }
 
     public void updateAuthor(AuthorDTO authorDTO) {
-        Author author = new Author();
-        author.setId(authorDTO.getId());
+        Author author = authorRepository.getAuthorById(authorDTO.getId());
+        if (author == null) {
+            throw new RuntimeException("Автор не найден"); // или бросить исключение
+        }
         author.setName(authorDTO.getName());
         author.setSurname(authorDTO.getSurname());
         authorRepository.updateAuthor(author);
     }
 
     public void deleteAuthor(Long id) {
+        Author author = authorRepository.getAuthorById(id);
+        if (author == null) {
+            throw new RuntimeException("Автор не найден"); // или бросить исключение
+        }
         authorRepository.deleteAuthor(id);
     }
 }
